# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PRISM Engine API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/questions/{track}": {
      "get": {
        "summary": "Questions",
        "operationId": "questions_api_questions__track__get",
        "parameters": [
          {
            "name": "track",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "school",
                "college"
              ],
              "type": "string",
              "title": "Track"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicQuestionSet"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/assess": {
      "post": {
        "summary": "Assess",
        "operationId": "assess_api_assess_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssessRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/whatif": {
      "post": {
        "summary": "Whatif",
        "operationId": "whatif_api_whatif_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WhatIfRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhatIfResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/explanations/{assessment_id}": {
      "get": {
        "summary": "Explanations",
        "operationId": "explanations_api_explanations__assessment_id__get",
        "parameters": [
          {
            "name": "assessment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Assessment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Explanations"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/careers/{career_id}": {
      "get": {
        "summary": "Career",
        "operationId": "career_api_careers__career_id__get",
        "parameters": [
          {
            "name": "career_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Career Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Career"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AssessRequest": {
        "properties": {
          "student": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/SchoolStudentInput"
              },
              {
                "$ref": "#/components/schemas/CollegeStudentInput"
              }
            ],
            "title": "Student",
            "discriminator": {
              "propertyName": "track",
              "mapping": {
                "college": "#/components/schemas/CollegeStudentInput",
                "school": "#/components/schemas/SchoolStudentInput"
              }
            }
          },
          "parent": {
            "$ref": "#/components/schemas/ParentInput"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "student",
          "parent"
        ],
        "title": "AssessRequest"
      },
      "AssessmentResult": {
        "properties": {
          "as_of": {
            "type": "string",
            "format": "date",
            "title": "As Of"
          },
          "ranking": {
            "items": {
              "$ref": "#/components/schemas/RankedCareer"
            },
            "type": "array",
            "title": "Ranking"
          },
          "details": {
            "additionalProperties": {
              "$ref": "#/components/schemas/CareerDetail"
            },
            "type": "object",
            "title": "Details"
          },
          "conflict": {
            "$ref": "#/components/schemas/ConflictResult"
          },
          "middle_path": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MiddlePath"
              },
              {
                "type": "null"
              }
            ]
          },
          "stretch_options": {
            "items": {
              "$ref": "#/components/schemas/StretchOption"
            },
            "type": "array",
            "title": "Stretch Options"
          },
          "alternatives": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DreamAlternatives"
              },
              {
                "type": "null"
              }
            ]
          },
          "swot": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Swot"
              },
              {
                "type": "null"
              }
            ]
          },
          "confidence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Confidence"
              },
              {
                "type": "null"
              }
            ]
          },
          "trace": {
            "$ref": "#/components/schemas/PipelineTrace"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "explanations": {
            "$ref": "#/components/schemas/Explanations"
          }
        },
        "type": "object",
        "required": [
          "as_of",
          "ranking",
          "details",
          "conflict",
          "middle_path",
          "stretch_options",
          "alternatives",
          "swot",
          "confidence",
          "trace",
          "id",
          "explanations"
        ],
        "title": "AssessmentResult"
      },
      "Career": {
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1,
            "title": "Id"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          },
          "domain": {
            "type": "string",
            "enum": [
              "Technology",
              "Engineering",
              "Science",
              "Health",
              "Arts & Design",
              "Finance & Maths",
              "Hyper-local"
            ],
            "title": "Domain"
          },
          "steam": {
            "items": {
              "type": "string",
              "enum": [
                "S",
                "T",
                "E",
                "A",
                "M"
              ]
            },
            "type": "array",
            "minItems": 1,
            "title": "Steam"
          },
          "summary": {
            "type": "string",
            "minLength": 1,
            "title": "Summary"
          },
          "requirement_vector": {
            "$ref": "#/components/schemas/RequirementVector"
          },
          "skills": {
            "items": {
              "$ref": "#/components/schemas/Skill"
            },
            "type": "array",
            "minItems": 1,
            "title": "Skills"
          },
          "pathways": {
            "items": {
              "$ref": "#/components/schemas/Pathway"
            },
            "type": "array",
            "maxItems": 4,
            "minItems": 2,
            "title": "Pathways"
          },
          "salary_inr_lpa": {
            "$ref": "#/components/schemas/SalaryBands"
          },
          "growth_index": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Growth Index"
          },
          "job_velocity": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Job Velocity"
          },
          "disruption_index": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Disruption Index"
          },
          "stability": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Stability"
          },
          "risk_level": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Risk Level"
          },
          "prestige": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Prestige"
          },
          "higher_studies_typical": {
            "type": "boolean",
            "title": "Higher Studies Typical"
          },
          "city_demand": {
            "additionalProperties": {
              "type": "number",
              "maximum": 1.0,
              "minimum": 0.0
            },
            "propertyNames": {
              "enum": [
                "bengaluru",
                "mysuru",
                "chennai",
                "hyderabad",
                "pune",
                "delhi_ncr",
                "mumbai",
                "coimbatore"
              ]
            },
            "type": "object",
            "title": "City Demand"
          },
          "adjacent_careers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Adjacent Careers"
          },
          "sources": {
            "items": {
              "type": "string",
              "minLength": 1
            },
            "type": "array",
            "minItems": 1,
            "title": "Sources"
          },
          "data_note": {
            "type": "string",
            "minLength": 1,
            "title": "Data Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "domain",
          "steam",
          "summary",
          "requirement_vector",
          "skills",
          "pathways",
          "salary_inr_lpa",
          "growth_index",
          "job_velocity",
          "disruption_index",
          "stability",
          "risk_level",
          "prestige",
          "higher_studies_typical",
          "city_demand",
          "adjacent_careers",
          "sources",
          "data_note"
        ],
        "title": "Career"
      },
...
      "CollegeStudentInput": {
        "properties": {
          "aptitude_answers": {
            "additionalProperties": {
              "type": "integer",
              "minimum": 0.0
            },
            "type": "object",
            "title": "Aptitude Answers",
            "default": {}
          },
          "riasec_answers": {
            "additionalProperties": {
              "type": "integer",
              "maximum": 5.0,
              "minimum": 1.0
            },
            "type": "object",
            "title": "Riasec Answers",
            "default": {}
          },
          "workstyle_answers": {
            "additionalProperties": {
              "type": "integer",
              "maximum": 5.0,
              "minimum": 1.0
            },
            "type": "object",
            "title": "Workstyle Answers",
            "default": {}
          },
          "marks_percent": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Marks Percent"
          },
          "favourite_subjects": {
            "items": {
              "type": "string",
              "minLength": 1
            },
            "type": "array",
            "title": "Favourite Subjects",
            "default": []
          },
          "home_city": {
            "type": "string",
            "enum": [
              "bengaluru",
              "mysuru",
              "chennai",
              "hyderabad",
              "pune",
              "delhi_ncr",
              "mumbai",
              "coimbatore"
            ],
            "title": "Home City"
          },
          "preferred_cities": {
            "items": {
              "type": "string",
              "enum": [
                "bengaluru",
                "mysuru",
                "chennai",
                "hyderabad",
                "pune",
                "delhi_ncr",
                "mumbai",
                "coimbatore"
              ]
            },
            "type": "array",
            "title": "Preferred Cities",
            "default": []
          },
          "willing_to_relocate": {
            "type": "boolean",
            "title": "Willing To Relocate"
          },
          "wants_higher_studies": {
            "type": "boolean",
            "title": "Wants Higher Studies"
          },
          "risk_tolerance": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "title": "Risk Tolerance"
          },
          "dream_career_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dream Career Id"
          },
          "free_text_1": {
            "type": "string",
            "title": "Free Text 1",
            "default": ""
          },
          "free_text_2": {
            "type": "string",
            "title": "Free Text 2",
            "default": ""
          },
          "track": {
            "type": "string",
            "const": "college",
            "title": "Track"
          },
          "degree": {
            "type": "string",
            "minLength": 1,
            "title": "Degree"
          },
          "year": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Year"
          },
          "self_rated_skills": {
            "additionalProperties": {
              "type": "integer",
              "maximum": 100.0,
              "minimum": 0.0
            },
            "type": "object",
            "title": "Self Rated Skills",
            "default": {}
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "home_city",
          "willing_to_relocate",
          "wants_higher_studies",
          "risk_tolerance",
          "track",
          "degree",
          "year"
        ],
        "title": "CollegeStudentInput"
      },
...
      "ConflictResult": {
        "properties": {
          "index": {
            "type": "number",
            "title": "Index"
          },
          "dimensions": {
            "items": {
              "$ref": "#/components/schemas/ConflictDimension"
            },
            "type": "array",
            "title": "Dimensions"
          },
          "hotspots": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Hotspots"
          },
          "student_top_domains": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Student Top Domains"
          }
        },
        "type": "object",
        "required": [
          "index",
          "dimensions",
          "hotspots",
          "student_top_domains"
        ],
        "title": "ConflictResult"
      },
...
      "Explanations": {
        "properties": {
          "source": {
            "type": "string",
            "enum": [
              "template",
              "llm"
            ],
            "title": "Source"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/CareerExplanation"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "source",
          "items"
        ],
        "title": "Explanations"
      },
      "FreeTextPrompt": {
        "properties": {
          "id": {
            "type": "string",
            "enum": [
              "free_text_1",
              "free_text_2"
            ],
            "title": "Id"
          },
          "prompt": {
            "type": "string",
            "minLength": 1,
            "title": "Prompt"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "prompt"
        ],
        "title": "FreeTextPrompt"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ParentInput": {
        "properties": {
          "annual_income_inr": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Annual Income Inr"
          },
          "education_budget_inr": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Education Budget Inr"
          },
          "loan_willingness": {
            "type": "string",
            "enum": [
              "none",
              "moderate",
              "high"
            ],
            "title": "Loan Willingness"
          },
          "risk_appetite": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "title": "Risk Appetite"
          },
          "preferred_domains": {
            "items": {
              "type": "string",
              "enum": [
                "Technology",
                "Engineering",
                "Science",
                "Health",
                "Arts & Design",
                "Finance & Maths",
                "Hyper-local"
              ]
            },
            "type": "array",
            "title": "Preferred Domains",
            "default": []
          },
          "location_preference": {
            "type": "string",
            "enum": [
              "near_home",
              "anywhere_india",
              "abroad_ok"
            ],
            "title": "Location Preference"
          },
          "supports_higher_studies": {
            "type": "boolean",
            "title": "Supports Higher Studies"
          },
          "top_priority": {
            "type": "string",
            "enum": [
              "stability",
              "salary",
              "prestige",
              "happiness"
            ],
            "title": "Top Priority"
          },
          "free_text": {
            "type": "string",
            "title": "Free Text",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "annual_income_inr",
          "education_budget_inr",
          "loan_willingness",
          "risk_appetite",
          "location_preference",
          "supports_higher_studies",
          "top_priority"
        ],
        "title": "ParentInput"
      },
...
      "PublicAptitudeItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "dimension": {
            "type": "string",
            "enum": [
              "numerical",
              "logical",
              "verbal",
              "spatial"
            ],
            "title": "Dimension"
          },
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "options": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Options"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "dimension",
          "prompt",
          "options"
        ],
        "title": "PublicAptitudeItem"
      },
      "PublicQuestionSet": {
        "properties": {
          "track": {
            "type": "string",
            "enum": [
              "school",
              "college"
            ],
            "title": "Track"
          },
          "aptitude": {
            "items": {
              "$ref": "#/components/schemas/PublicAptitudeItem"
            },
            "type": "array",
            "title": "Aptitude"
          },
          "riasec": {
            "items": {
              "$ref": "#/components/schemas/RiasecItem"
            },
            "type": "array",
            "title": "Riasec"
          },
          "workstyle": {
            "items": {
              "$ref": "#/components/schemas/WorkstyleItem"
            },
            "type": "array",
            "title": "Workstyle"
          },
          "free_text": {
            "items": {
              "$ref": "#/components/schemas/FreeTextPrompt"
            },
            "type": "array",
            "title": "Free Text"
          },
          "skills": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skills"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "track",
          "aptitude",
          "riasec",
          "workstyle",
          "free_text"
        ],
        "title": "PublicQuestionSet",
        "description": "\u00a713 question set as sent to the browser: aptitude answer keys stay on the server."
      },
...
      "RiasecItem": {
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1,
            "title": "Id"
          },
          "dimension": {
            "type": "string",
            "enum": [
              "R",
              "I",
              "A",
              "S",
              "E",
              "C"
            ],
            "title": "Dimension"
          },
          "prompt": {
            "type": "string",
            "minLength": 1,
            "title": "Prompt"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "dimension",
          "prompt"
        ],
        "title": "RiasecItem"
      },
...
      "SchoolStudentInput": {
        "properties": {
          "aptitude_answers": {
            "additionalProperties": {
              "type": "integer",
              "minimum": 0.0
            },
            "type": "object",
            "title": "Aptitude Answers",
            "default": {}
          },
          "riasec_answers": {
            "additionalProperties": {
              "type": "integer",
              "maximum": 5.0,
              "minimum": 1.0
            },
            "type": "object",
            "title": "Riasec Answers",
            "default": {}
          },
          "workstyle_answers": {
            "additionalProperties": {
              "type": "integer",
              "maximum": 5.0,
              "minimum": 1.0
            },
            "type": "object",
            "title": "Workstyle Answers",
            "default": {}
          },
          "marks_percent": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Marks Percent"
          },
          "favourite_subjects": {
            "items": {
              "type": "string",
              "minLength": 1
            },
            "type": "array",
            "title": "Favourite Subjects",
            "default": []
          },
          "home_city": {
            "type": "string",
            "enum": [
              "bengaluru",
              "mysuru",
              "chennai",
              "hyderabad",
              "pune",
              "delhi_ncr",
              "mumbai",
              "coimbatore"
            ],
            "title": "Home City"
          },
          "preferred_cities": {
            "items": {
              "type": "string",
              "enum": [
                "bengaluru",
                "mysuru",
                "chennai",
                "hyderabad",
                "pune",
                "delhi_ncr",
                "mumbai",
                "coimbatore"
              ]
            },
            "type": "array",
            "title": "Preferred Cities",
            "default": []
          },
          "willing_to_relocate": {
            "type": "boolean",
            "title": "Willing To Relocate"
          },
          "wants_higher_studies": {
            "type": "boolean",
            "title": "Wants Higher Studies"
          },
          "risk_tolerance": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "title": "Risk Tolerance"
          },
          "dream_career_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dream Career Id"
          },
          "free_text_1": {
            "type": "string",
            "title": "Free Text 1",
            "default": ""
          },
          "free_text_2": {
            "type": "string",
            "title": "Free Text 2",
            "default": ""
          },
          "track": {
            "type": "string",
            "const": "school",
            "title": "Track"
          },
          "current_class": {
            "type": "integer",
            "maximum": 12.0,
            "minimum": 9.0,
            "title": "Current Class"
          },
          "stream": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "PCM",
                  "PCB",
                  "PCMB",
                  "Commerce",
                  "Humanities"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Stream"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "home_city",
          "willing_to_relocate",
          "wants_higher_studies",
          "risk_tolerance",
          "track",
          "current_class"
        ],
        "title": "SchoolStudentInput"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WhatIfRequest": {
        "properties": {
          "assessment_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assessment Id"
          },
          "profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssessRequest"
              },
              {
                "type": "null"
              }
            ]
          },
          "overrides": {
            "$ref": "#/components/schemas/WhatIfOverrides",
            "default": {}
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "WhatIfRequest"
      },
      "WhatIfResult": {
        "properties": {
          "assessment_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assessment Id"
          },
          "overrides": {
            "$ref": "#/components/schemas/WhatIfOverrides"
          },
          "ranking": {
            "items": {
              "$ref": "#/components/schemas/RankedCareer"
            },
            "type": "array",
            "title": "Ranking"
          },
          "changes": {
            "items": {
              "$ref": "#/components/schemas/CareerChange"
            },
            "type": "array",
            "title": "Changes"
          },
          "elapsed_ms": {
            "type": "number",
            "title": "Elapsed Ms"
          }
        },
        "type": "object",
        "required": [
          "assessment_id",
          "overrides",
          "ranking",
          "changes",
          "elapsed_ms"
        ],
        "title": "WhatIfResult"
      },
      "WorkstyleItem": {
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1,
            "title": "Id"
          },
          "dimension": {
            "type": "string",
            "enum": [
              "creative",
              "stability_vs_excitement",
              "teamwork",
              "structure"
            ],
            "title": "Dimension"
          },
          "prompt": {
            "type": "string",
            "minLength": 1,
            "title": "Prompt"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "dimension",
          "prompt"
        ],
        "title": "WorkstyleItem"
      }
    }
  }
}